requirements.txt
numpy
pandas

transformers
torch

flask

python-dotenv

torchvision
pillow